+++ site/homepage/content/release-notes.md
@@ -35,6 +35,10 @@ description = "Information about changes in recent Hono releases. Includes new f
 * The `keyStorePassword` and `trustStorePassword` properties of the Hono clients now also support specifying a file
   to read the password from. If the property value starts with `file:` then the value after the prefix is interpreted as
   as the path to a file to read the password from.
+* The registry implementations failed to reject a request to update a device's empty set of credentials, e.g. right
+  after the device has been created, if the request contained a secret having an ID. In fact, the registry
+  implementations would have accepted such a request even if the secrets did not contain a password/key/etc at all
+  but only the identifier. This has been fixed.
 
 ### Deprecations